moving towards float heights but error metric needs adjusting

+++ quadmerge/uquadtree.hpp
@@ -72,7 +72,7 @@ struct HeightMapInfo {
 
 struct	VertInfo {
   //	uint16	Z;
-  int Z;
+  float  Z;
   unsigned char num_samples;
   float *Zsamples;
 	unsigned char	Lightness;	// For simple precomputed vertex lighting for purposes of the demo.  It's a waste of 2 bytes if we're texturing.
